feat: move fmt implementations in a different module
